feature/detail-activity-addition Refactor: Clean current code
Delete unused code and remove the initial store of data
