+++ app/src/main/java/com/architectcoders/arquitectomarvel/model/database/ItemComics.kt
@@ -1,4 +1,4 @@
-package com.architectcoders.arquitectomarvel.model.database.character
+package com.architectcoders.arquitectomarvel.model.database
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
