+++ app/src/main/java/com/architectcoders/arquitectomarvel/model/database/relations/ResultWithItemsComics.kt
@@ -1,22 +1,20 @@
-package com.architectcoders.arquitectomarvel.model.database.character.relations
+package com.architectcoders.arquitectomarvel.model.database.relations
 
 import androidx.room.Embedded
 import androidx.room.Relation
 import com.architectcoders.arquitectomarvel.model.characters.Comics
 import com.architectcoders.arquitectomarvel.model.characters.Result
-import com.architectcoders.arquitectomarvel.model.database.character.ItemComics
-import com.architectcoders.arquitectomarvel.model.database.character.ResultRoom
-import com.architectcoders.arquitectomarvel.model.database.character.toItems
+import com.architectcoders.arquitectomarvel.model.database.ItemComics
+import com.architectcoders.arquitectomarvel.model.database.ResultRoom
+import com.architectcoders.arquitectomarvel.model.database.toItems
 
 data class ResultWithItemsComics(
-
     @Embedded val resultRoom: ResultRoom,
     @Relation(
         parentColumn = "comicsCollectionURI",
         entityColumn = "collectionURI"
     )
     val itemsComics: List<ItemComics>
-
 )
 
 val ResultWithItemsComics.toRemoteResult: Result
