--- app/src/main/java/com/architectcoders/arquitectomarvel/ui/main/ClickListener.kt
@@ -1,7 +1,0 @@
-package com.architectcoders.arquitectomarvel.ui.main
-
-import com.architectcoders.arquitectomarvel.model.characters.Result
-
-interface ClickListener {
-    fun onClick(mediaService: Result)
-}
